[move/#45] feedback 관련 화면 위치 변경
