+++ shared/feature/my/src/commonMain/kotlin/com/kus/feature/my/di/myFeatureModule.kt
@@ -7,7 +7,7 @@ import com.kus.feature.my.ui.community.MyScrapViewModel
 import com.kus.feature.my.ui.editprofile.EditProfileViewModel
 import com.kus.feature.my.ui.restaurant.CheckedResViewModel
 import com.kus.feature.my.ui.restaurant.FavoriteResViewModel
-import com.kus.feature.my.ui.restaurant.FeedbackViewModel
+import com.kus.feature.my.ui.feedback.FeedbackViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
