+++ shared/feature/my/src/commonMain/kotlin/com/kus/feature/my/ui/feedback/FeedbackViewModel.kt
@@ -1,4 +1,4 @@
-package com.kus.feature.my.ui.restaurant
+package com.kus.feature.my.ui.feedback
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
